Fix double white space error and small tidy of race class
Fix the error where if a driver had a double white space in their name then they
would be entered into the race class results as another driver (but with the
same name).
Added checkbox for this, checked by default.

Tidied the race class, instead of keeping a record of all drivers as they are
added, we now just iterate over the events when needed.